Working on getting Deep Open Classification
I am working on trying to get deep open classification to work. I am also moving the Loop over unknowns function into the Config so that we can just refrence the value instead of recalculating it. Makes it easier for DOC to find it.

Currently testing to see if this works.