Data-loaders+ and tensorboard
Added new Data-loaders and added some tensorboard integration. This has required rewriting some things in almost all of the files for one or both of these changes. (oh, and measurement is now an object not a function, seemed easier)
One more thing, also changed the training tensor for weibul fitting to not first be transferred to the GPU because the calculations need the CPU and it was just being transferred back