Fixed the weird CUDA error
It was caused by feeding a non CUDA tensor into the model (main.py line 114). Why it only happened on LOOP==3 and Looplevel ==11 is still a mystery but it should not happen any more.

also added a limit on evaluation, it needs at least 2 items per batch otherwise one of the dimensions collapses. And the code does not know how to handle that.