Formatting and 3 configs
Changed the formatting on one of the files because I installed Flake8 which shows formatting errors. But there are too many so I disabled it for now.

Also, 3 new config options:
AttemptLoad has been split into Attempt load Model and Attempt Load Data.
And ApplyPrelimSoft has been added which adds a preliminary softmax function to the model and only applies the unknown detection if the softmax dips below the threshold percentage.